search: extract shared simulated-MC test helpers to conftest.py
